# Improvements to open the addon faster