### Check if this PR fulfills these requirements:
* [x] My changes respect the [Kodi add-on development rules](https://kodi.wiki/view/Add-on_rules)
* [x] I have read the [**CONTRIBUTING**](Contributing.md) document.
* [x] I made sure there wasn't another one [Pull Requests](../../pulls) opened for the same update/change
* [x] I have successfully tested my changes locally

#### Types of changes
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Feature change (non-breaking change which change behaviour of an existing functionality)
- [x] Improvement (non-breaking change which improve functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

### Description
<!--- Motivation and a possible detailed explanation of the changes -->
<!--- Put your text below this line -->
Some time ago i have found the path request to get the profiles list from the API,
was hidden among the Netflix html answers in the PostGateData.
This allows you to safely obtain the list of profiles avoiding parsing transformations, 
and also avoid parserizing the web page every time you return to the profiles screen,
this allow to open the profiles screen in about only 1 second! instead of 3 seconds or more!

I also noticed that the requests module takes from 2 to 3 seconds longer than normal to process the first call, this freezes the opening of the addon for a very long time,
so to camouflage this problem, ~i preferred to force the update of the session data~ an update of the session headers will be performed when the service is started this way you can recover about 2 seconds!

~Can be configured in expert settings, will make happy those who use only kodi library and/or widgets~

Execution time test before (at first run of the addon):
make_call                      3219 ms
Execution time test after (at first run of the addon):
make_call                      1310 ms

PS. the "seconds" specified above, are based on tests performed on Windows 10, with fairly high network latency compared to others, so in my configuration the result data (the seconds) are more accentuated

### In case of Feature change / Breaking change:
#### Describe the current behavior
<!--- Put your text below this line -->

#### Describe the new behavior
<!--- Put your text below this line -->

### Screenshots (if appropriate):
<!--- Add some screenshots if they can be useful to show the differences between before and after the changes -->