@CastagnaIT 2020-01-07
Hi @dagwieers can you please test this on a RPI ?
or do you know someone who can?

i need to know if there are any side effects/errors to the first opening of the frontend of the addon

I basically need to verify this:
1. Install this addon branch version
2. Open expert settings and set Debug to verbose, and, Enable execution timing to ON
3. Reboot Kodi (do not use disable/enable addon, but always reboot, if you do more that one test)

Now, IMPORTANT: open the frontend immediately (if you can at the same time) when you see the notification of the started service (in this way should cause a double request via requests module before the previous one is actually finished)

if no errors occur, should be opening the profiles screen, so check the log file,
and looking for a line like this, towards the end of the file:
<pre>
2020-01-07 15:12:20.017 T:1652   DEBUG: [plugin.video.netflix (1)] Execution time info for this run:
                                            build_profiles_listing           17 ms
                                            make_call                       805 ms
</pre>
compared to the version in the master branch should result in a value theoretically approximately halved