@dagwieers 2020-01-07
@CastagnaIT 

I did a quick test on LE 9.2.0 (development branch, it's Kodi 18.5 with some known fixes) and it worked. But it's hard to find the correct timings in the logs with all the stuff that is being done in the background and Kodi internals.

Original v0.16.1:
```
2020-01-07 18:43:33.852 T:1375978368   DEBUG: [plugin.video.netflix (1)] Execution time info for this run:
                                            build_profiles_listing           66 ms
                                            make_call                       573 ms
2020-01-07 18:46:19.153 T:1388299136   DEBUG: [plugin.video.netflix (1)] Execution time info for this run:
                                            build_profiles_listing           71 ms
                                            make_call                       503 ms
2020-01-07 18:47:58.806 T:1378857856   DEBUG: [plugin.video.netflix (1)] Execution time info for this run:
                                            build_profiles_listing           79 ms
                                            make_call                       523 ms
```

Using this PR:
```
2020-01-07 18:40:06.992 T:1376768896   DEBUG: [plugin.video.netflix (1)] Execution time info for this run:
                                            build_profiles_listing           38 ms
                                            make_call                       137 ms
2020-01-07 18:50:00.850 T:1388827520   DEBUG: [plugin.video.netflix (1)] Execution time info for this run:
                                            build_profiles_listing           38 ms
                                            make_call                       168 ms
2020-01-07 18:51:40.661 T:1385161600   DEBUG: [plugin.video.netflix (1)] Execution time info for this run:
                                            build_profiles_listing           84 ms
                                            make_call                       193 ms
2020-01-07 18:53:29.933 T:1396417408   DEBUG: [plugin.video.netflix (1)] Execution time info for this run:
                                            build_profiles_listing           63 ms
                                            make_call                       169 ms
```

Startup takes about 3.5 turns on the spinner (about 2 seconds?). It doesn't appear to make a lot of difference to perceived startup time (wrt. the spinner) but that is probably not a good metric anyway ;-)

PS I assume disable/enable doesn't help because of window property caching or something similar? Maybe it could be useful for testing to have the option to purge any caching or other metrics-influencing functionality?