@CastagnaIT 2020-01-07
thank you!
i say don't use "disable/enable" to have a complete cold situation
this nasty hack works great! amazing
in my computer you can notice much more i have more latency in my internet connection